util/type_traits.hpp renamed to common.hpp
