+++ cpp/oneapi/dal/util/common.hpp
